+++ tests/test-661av1_in_matroska_by_mkvmerge_28_0_0.rb
@@ -1,0 +1,5 @@
+#!/usr/bin/ruby -w
+
+# T_661av1_in_matroska_by_mkvmerge_28_0_0
+describe "mkvmerge / AV1 in Matroska created by mkvmerge v28.0.0"
+test_merge "data/mkv/av1-mkvmerge-28.0.0.mkv"